Update patch set 14
Patch Set 14:

(16 comments)

I have left some minor comments.

The only change I would strongly encourage is to move the _get_message_data function out of run, into utils (or maybe run_handler), and unit test it. This function should be unit tested and the bokeh file isn't. 

The rest of the comments are small and could be addressed in a subsequent PR as this PR is already large enough.

Patch-set: 14
Label: Code-Review=0