Update patch set 1
Patch Set 1:

> This is a Python project .. thought Sonar didn't work on that?

Yes, Sonar works with Python projects. This is for RELENG-1230 (apologies for not citing that in the commit message), which is based on RT ticket 61030, where Manoop requested Python sonar scans similar to those on ONAP. In order to get those working, we will have to modify or add pom.xml files to the Python projects.

Patch-set: 1
Reviewer: Gerrit User 223 <223@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=0