Update patch set 1
Patch Set 1:

I will submit this because it's the way the build process does things, but the fact that there is a poml.xml file in a *python* project is still offensive to me. Rather than going further down that path, I would have preferred this sonar change to be in like a sonar.properties file or something along those lines. Now these poms are clearly here 4 life.

Patch-set: 1