Update patch set 2
Patch Set 2:

Setup is for abstract requirements. Requirements are for hard requirements.  

One thing that is common in setup though is some kind of version ranging, for example we could say bokeh<2.0.0. 

If you have a requirement that one of these packages get pinned, or ranged, let me know, beyond general unspecified unease. This is the way people do things. Yes at some point it will break but you want to find those situations. 

RE tox, many people may not be on 3.7 yet, it is generally good practice to support all python versions that work if there is no overhead in doing so, and in this case there is none. If supporting an older version required work in the code, I would agree. As an example see Paul's client lib tox, he even has special tests for specific python versions. 

RE the version, OK I'll make the change, but that breaks semantic versioning, as there was no API change so this should definitely not bump to 2.0.0 out of correctness.

Patch-set: 2