Update patch set 4
Patch Set 4:

OK I thought about this more and I guess underpinning and overpinning the major versions seems like an ok thing to do. 

I still feel that it is over-restrictive, but there isn't really a good reason for someone to use bokeh 0 now that 1 is known and working (better), so it seems reasonable to not allow <1, etc.

Patch-set: 4
Label: Code-Review=0